pyproject.toml
[project]
name = "mcp-local"
version = "0.1.0"
description = "MCP Starterprojekt mit automatischer Unit-Registrierung"
authors = [
  { name = "DATA-MINT-RESEARCH", email = "data.mint.research@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"

dependencies = [
  "pyyaml>=6.0",
  "docker>=6.1.0",
  "docker-compose>=1.29.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "coverage>=7.0",
  "pytest-cov>=4.0",
  "docker>=6.1.0",
  "docker-compose>=1.29.2"
]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"